replaced long double with double, replaced pointers with references, added arduino sketch

+++ software/motioncalc/Matrix3D.cpp
@@ -10,7 +10,7 @@ Matrix3D::Matrix3D()
 	}
 }
 
-Matrix3D::Matrix3D(long double val[3][3])
+Matrix3D::Matrix3D(double val[3][3])
 {
 	for (int i = 0; i < 3; i++) {
 		for (int j = 0; j < 3; j++) {
@@ -45,18 +45,18 @@ Matrix3D & Matrix3D::operator*=(const Matrix3D & y)
 	return *this;
 }
 
-Matrix3D Matrix3D::fromRotation(int axis, long double angle)
+Matrix3D Matrix3D::fromRotation(int axis, double angle)
 {
 	if (axis == X_AXIS) {
 		
-		long double values[3][3] = { { 1, 0, 0 }, { 0, cosl(angle), -sinl(angle) }, {0, sinl(angle), cosl(angle)} };
+		double values[3][3] = { { 1, 0, 0 }, { 0, cos(angle), -sin(angle) }, {0, sin(angle), cos(angle)} };
 		return Matrix3D(values);
 	}
 	else if (axis == Y_AXIS) {
-		long double values[3][3] = { { cosl(angle), 0, sinl(angle) },{ 0, 1, 0},{ -sinl(angle), 0, cosl(angle) } };
+		double values[3][3] = { { cos(angle), 0, sin(angle) },{ 0, 1, 0},{ -sin(angle), 0, cos(angle) } };
 		return Matrix3D(values);
 	}
-	long double values[3][3] = { { cosl(angle), -sinl(angle), 0},{ sinl(angle), cosl(angle), 0 },{ 0, 0, 1 } };
+	double values[3][3] = { { cos(angle), -sin(angle), 0},{ sin(angle), cos(angle), 0 },{ 0, 0, 1 } };
 	return Matrix3D(values);
 }
 	
@@ -83,10 +83,10 @@ inline Matrix3D operator+(Matrix3D x, const Matrix3D & y)
 
 inline Matrix3D operator*(Matrix3D x, const Matrix3D & y)
 {
-	long double values[3][3];
+	double values[3][3];
 	for (int i = 0; i < 3; i++) {
 		for (int j = 0; j < 3; j++) {
-			long double sum = 0;
+			double sum = 0;
 			for (int k = 0; k < 3; k++) {
 				sum += x.values[i][k] * y.values[k][j];
 			}
